back spring trading: added business logic for updating portfolio units qty and moneypool unit price when add/withdraw from portfolio or completed trades

+++ back/spring/src/main/java/com/robotrader/spring/model/MoneyPool.java
@@ -40,7 +40,7 @@ public class MoneyPool {
 
     @NotNull(message = "Unit quantity cannot be null")
     @Min(value = 0, message = "Unit quantity must be greater than or equal to 0")
-    private Integer totalUnitQty;
+    private BigDecimal totalUnitQty;
 
     @NotNull
     @Temporal(TemporalType.TIMESTAMP)
